Addition of Spot Lights, not working yet though

+++ PointLight.cpp
@@ -20,9 +20,9 @@ PointLight::PointLight(GLfloat red, GLfloat green, GLfloat blue,
     exponent = exp;
 }
 
-void PointLight::UseLight(GLfloat ambientIntensityLocation, GLfloat ambientColourLocation,
-                          GLfloat diffuseIntensityLocation, GLfloat positionLocation,
-                          GLfloat constantLocation, GLfloat linearLocation, GLfloat exponentLocation)
+void PointLight::UseLight(GLuint ambientIntensityLocation, GLuint ambientColourLocation,
+                          GLuint diffuseIntensityLocation, GLuint positionLocation,
+                          GLuint constantLocation, GLuint linearLocation, GLuint exponentLocation)
 {
     glUniform3f(ambientColourLocation, colour.x, colour.y, colour.z);
     glUniform1f(ambientIntensityLocation, ambientIntensity);
